Добавлены ограничения на методы, кроме getAllContacts и getContactById

+++ src/main/java/com/example/MyPersonalContactManager/controller/ContactController.java
@@ -2,8 +2,10 @@ package com.example.MyPersonalContactManager.controller;
 
 import com.example.MyPersonalContactManager.models.ContactModels.Contact;
 import com.example.MyPersonalContactManager.models.ContactModels.ContactDTOBig;
+import com.example.MyPersonalContactManager.models.Error;
 import com.example.MyPersonalContactManager.models.RequestResponseModels.RequestBodyClient;
 import com.example.MyPersonalContactManager.models.RequestResponseModels.ResponseAPI;
+import com.example.MyPersonalContactManager.repository.DatabaseUserRepository;
 import com.example.MyPersonalContactManager.service.DatabaseContactService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,16 +19,28 @@ public class ContactController {
     private ResponseAPI responseAPI;
 
     @Autowired
+    private DatabaseUserRepository dbUserRepository;
+    @Autowired
     private DatabaseContactService dbService;
 
-
     @PostMapping(value = "/createContact", consumes = "application/json")
-    public ResponseEntity<ResponseAPI> crateContact(@Valid @RequestBody RequestBodyClient requestBodyClient) {
-        responseAPI = new ResponseAPI();
+    public ResponseEntity<ResponseAPI> crateContact(@RequestHeader("Authorization") String token,
+                                                    @Valid @RequestBody RequestBodyClient requestBodyClient) {
+        if (token == null || token.isEmpty()) {
+            responseAPI.response = new Error(400, "Authorization header is missing.");
+            return ResponseEntity.badRequest().body(responseAPI);
+        }
 
-        Contact contact = dbService.createContact(requestBodyClient.contact);
-        responseAPI.response = contact;
-        return ResponseEntity.ok(responseAPI);
+        boolean userRole = dbUserRepository.getUserRoleByToken(token);
+        responseAPI = new ResponseAPI();
+        if (userRole) {
+            Contact contact = dbService.createContact(requestBodyClient.contact);
+            responseAPI.response = contact;
+            return ResponseEntity.ok(responseAPI);
+        } else {
+            responseAPI.response = new Error(403, "Access denied.");
+            return ResponseEntity.ok(responseAPI);
+        }
     }
 
     @GetMapping("/contacts/{id}")
@@ -44,17 +58,45 @@ public class ContactController {
         return ResponseEntity.ok(responseAPI);
     }
 
-    @PutMapping(value = "/updateContact")
-    public ResponseEntity<ResponseAPI> updateContact(@PathVariable String id, @RequestBody RequestBodyClient requestBodyClient) {
+    @PutMapping(value = "/updateContact/{id}", consumes = "application/json")
+    public ResponseEntity<ResponseAPI> updateContact(@PathVariable String id,
+                                                     @RequestHeader("Authorization") String token,
+                                                     @RequestBody RequestBodyClient requestBodyClient) {
+        if (token == null || token.isEmpty()) {
+            responseAPI.response = new Error(400, "Authorization header is missing.");
+            return ResponseEntity.badRequest().body(responseAPI);
+        }
+
+        boolean userRole = dbUserRepository.getUserRoleByToken(token);
         responseAPI = new ResponseAPI();
-        ContactDTOBig contactDTOBig = dbService.updateContact(id, requestBodyClient.contactDTOBig);
-        responseAPI.response = contactDTOBig;
-        return ResponseEntity.ok(responseAPI);
+        if (userRole) {
+            ContactDTOBig contactDTOBig = dbService.updateContact(id, requestBodyClient.contactDTOBig);
+            responseAPI.response = contactDTOBig;
+            return ResponseEntity.ok(responseAPI);
+        } else {
+            responseAPI.response = new Error(403, "Access denied.");
+            return ResponseEntity.ok(responseAPI);
+        }
     }
 
-    @DeleteMapping("contacts/{id}")
-    public ResponseEntity<Boolean> deleteContactById(@PathVariable String id) {
-        boolean isDeleted = dbService.deleteContactById(id);
-        return ResponseEntity.ok(isDeleted);
+    @DeleteMapping("contacts/delete")
+    public ResponseEntity<ResponseAPI> deleteContactById(@RequestHeader("Contact-Id") String id,
+                                                         @RequestHeader("Authorization") String token) {
+        if (token == null || token.isEmpty()) {
+            responseAPI.response = new Error(400, "Authorization header is missing.");
+            return ResponseEntity.badRequest().body(responseAPI);
+        }
+
+        boolean userRole = dbUserRepository.getUserRoleByToken(token);
+        responseAPI = new ResponseAPI();
+        boolean isDeleted;
+        if (userRole) {
+            isDeleted = dbService.deleteContactById(id);
+            responseAPI.response = isDeleted;
+            return ResponseEntity.ok(responseAPI);
+        } else {
+            responseAPI.response = new Error(403, "Access denied.");
+            return ResponseEntity.ok(responseAPI);
+        }
     }
 }
